further decoupling + naming change
