+++ src/ResolverUtils.ts
@@ -2,7 +2,12 @@ import isPromise from 'is-promise';
 import { setImmediate } from 'tiny-set-immediate';
 import warning from 'tiny-warning';
 
-import { Match, RouteIndices, RouteMatch, RouteObjectBase } from './typeUtils';
+import {
+  Match,
+  RouteIndices,
+  RouteMatch,
+  RouteObjectBase,
+} from './utilityTypes';
 
 const UNRESOLVED = {};
