+++ src/useMatch.ts
@@ -1,5 +1,5 @@
-import { Match } from './typeUtils';
 import useRouter from './useRouter';
+import { Match } from './utilityTypes';
 
 /** Returns the current route Match */
 export default function useMatch<TContext = any>(): Match<TContext> | null {
