+++ src/useParams.ts
@@ -1,5 +1,5 @@
-import { Params } from './typeUtils';
 import useMatch from './useMatch';
+import { Params } from './utilityTypes';
 
 /** Returns the current route params */
 export default function useParams(): Params | undefined {
